fix migrate from supabase to pocketbase

+++ lib/app/data/repositories/booking_repository.dart
@@ -11,7 +11,7 @@ class BookingRepository {
   Future<void> createBooking(BookingModel booking) async {
     await _database.create(
       collection: collection,
-      documentId: booking.id!,
+      documentId: booking.id ?? '',
       data: booking.toJson(),
     );
   }
@@ -19,10 +19,16 @@ class BookingRepository {
   Future<List<BookingModel>> getBookings() async {
     final result = await _database.readAll(collection: collection);
 
-    return result.map((item) => BookingModel.fromJson(item)).toList();
+    return result.map((item) {
+      return BookingModel.fromJson(item);
+    }).toList();
   }
 
   Future<void> updateBooking(BookingModel booking) async {
+    if (booking.id == null || booking.id!.isEmpty) {
+      throw Exception('Booking ID tidak boleh kosong saat update');
+    }
+
     await _database.update(
       collection: collection,
       documentId: booking.id!,
@@ -31,6 +37,10 @@ class BookingRepository {
   }
 
   Future<void> deleteBooking(String bookingId) async {
+    if (bookingId.isEmpty) {
+      throw Exception('Booking ID tidak boleh kosong saat delete');
+    }
+
     await _database.delete(collection: collection, documentId: bookingId);
   }
 }
