+++ lib/app/data/repositories/package_repository.dart
@@ -11,17 +11,24 @@ class PackageRepository {
   Future<void> createPackage(PackageModel package) async {
     await _database.create(
       collection: collection,
-      documentId: package.id!,
+      documentId: package.id ?? '',
       data: package.toJson(),
     );
   }
 
   Future<List<PackageModel>> getPackages() async {
     final result = await _database.readAll(collection: collection);
-    return result.map((item) => PackageModel.fromJson(item)).toList();
+
+    return result.map((item) {
+      return PackageModel.fromJson(item);
+    }).toList();
   }
 
   Future<void> updatePackage(PackageModel package) async {
+    if (package.id == null || package.id!.isEmpty) {
+      throw Exception('Package ID tidak boleh kosong saat update');
+    }
+
     await _database.update(
       collection: collection,
       documentId: package.id!,
@@ -30,6 +37,10 @@ class PackageRepository {
   }
 
   Future<void> deletePackage(String packageId) async {
+    if (packageId.isEmpty) {
+      throw Exception('Package ID tidak boleh kosong saat delete');
+    }
+
     await _database.delete(collection: collection, documentId: packageId);
   }
 }
